Code cleanup & linting enforcement
