+++ src/python/experiment_platform/converters.py
@@ -1,9 +1,14 @@
-import pandas as pd
+"""
+Collection of functions to handle data cleanup and conversion
+"""
 import numpy as np
 from sklearn import preprocessing
 
 
 def dates_to_floats(X, y):
+    """
+    Convert datetimes to floats of seconds since epoch
+    """
     dates = X.select_dtypes(
         include=["datetime64", "datetime64[ns]", "datetime64[ns, UTC]"]
     )
@@ -13,6 +18,9 @@ def dates_to_floats(X, y):
 
 
 def binarize_y(X, y):
+    """
+    Binarizes labels
+    """
     if y.dtype == object:
         lb = preprocessing.LabelBinarizer()
         y = lb.fit_transform(y).ravel()
@@ -20,6 +28,9 @@ def binarize_y(X, y):
 
 
 def replace_nan(X, y):
+    """
+    Replace NaNs in float columns in place with 0s
+    """
     floats = X.select_dtypes(include=["float32", "float64"])
     for col in floats.columns:
         X[col] = X[col].replace(np.nan, 0)
@@ -27,10 +38,13 @@ def replace_nan(X, y):
 
 
 def drop_objects(X, y):
+    """
+    Drop all object type columns
+    """
     return X.select_dtypes(exclude=["object"]), y
 
 
-all_converters = [
+ALL_CONVERTERS = [
     dates_to_floats,
     replace_nan,
     binarize_y,
